Changed longs to Integers in Facilities for consistent number types;

+++ src/main/java/Facilities/AtlantaFacility.java
@@ -13,8 +13,8 @@ import java.util.*;
 public final class AtlantaFacility implements Facility, Comparable<Facility>{
 
     private String location = "Atlanta, GA";
-    private long ratePerDay = 10;
-    private long costPerDay = 300;
+    private Integer ratePerDay = 10;
+    private Integer costPerDay = 300;
     private ArrayList<FacilityEdge> neighbors = new ArrayList<>();
 
     private HashMap<Item, Integer> activeInventory = new HashMap<>();
@@ -54,7 +54,7 @@ public final class AtlantaFacility implements Facility, Comparable<Facility>{
     }
 
     @Override
-    public Long getRatePerDay() {
+    public Integer getRatePerDay() {
         return ratePerDay;
     }
 
@@ -64,7 +64,7 @@ public final class AtlantaFacility implements Facility, Comparable<Facility>{
     }
 
     @Override
-    public Long getCostPerDay() {
+    public Integer getCostPerDay() {
         return costPerDay;
     }
