+++ src/main/java/Facilities/BostonFacility.java
@@ -17,8 +17,8 @@ public final class BostonFacility implements Facility, Comparable<Facility> {
 
     private String location = "Boston, MA";
 
-    private long ratePerDay = 10;
-    private long costPerDay = 300;
+    private Integer ratePerDay = 10;
+    private Integer costPerDay = 300;
     private ArrayList<FacilityEdge> neighbors = new ArrayList<>();
 
     private HashMap<Item, Integer> activeInventory = new HashMap<>();
@@ -57,7 +57,7 @@ public final class BostonFacility implements Facility, Comparable<Facility> {
     }
 
     @Override
-    public Long getRatePerDay() {
+    public Integer getRatePerDay() {
         return ratePerDay;
     }
 
@@ -67,7 +67,7 @@ public final class BostonFacility implements Facility, Comparable<Facility> {
     }
 
     @Override
-    public Long getCostPerDay() {
+    public Integer getCostPerDay() {
         return costPerDay;
     }
