+++ src/main/java/Facilities/DenverFacility.java
@@ -16,8 +16,8 @@ import java.util.Map;
 public final class DenverFacility implements Facility, Comparable<Facility> {
 
     private String location = "Denver, CO";
-    private long ratePerDay = 10;
-    private long costPerDay = 300;
+    private Integer ratePerDay = 10;
+    private Integer costPerDay = 300;
 
     private ArrayList<FacilityEdge> neighbors = new ArrayList<>();
 
@@ -58,7 +58,7 @@ public final class DenverFacility implements Facility, Comparable<Facility> {
     }
 
     @Override
-    public Long getRatePerDay() {
+    public Integer getRatePerDay() {
         return ratePerDay;
     }
 
@@ -68,7 +68,7 @@ public final class DenverFacility implements Facility, Comparable<Facility> {
     }
 
     @Override
-    public Long getCostPerDay() {
+    public Integer getCostPerDay() {
         return costPerDay;
     }
