+++ src/main/java/Facilities/PhoenixFacility.java
@@ -17,8 +17,8 @@ import java.util.Map;
 public final class PhoenixFacility implements Facility, Comparable<Facility> {
 
     private String location = "Phoenix, AZ";
-    private long ratePerDay = 8;
-    private long costPerDay = 300;
+    private Integer ratePerDay = 8;
+    private Integer costPerDay = 300;
     private ArrayList<FacilityEdge> neighbors = new ArrayList<>();
 
     private HashMap<Item, Integer> activeInventory = new HashMap<>();
@@ -59,7 +59,7 @@ public final class PhoenixFacility implements Facility, Comparable<Facility> {
     }
 
     @Override
-    public Long getRatePerDay() {
+    public Integer getRatePerDay() {
         return ratePerDay;
     }
 
@@ -69,7 +69,7 @@ public final class PhoenixFacility implements Facility, Comparable<Facility> {
     }
 
     @Override
-    public Long getCostPerDay() {
+    public Integer getCostPerDay() {
         return costPerDay;
     }
