+++ src/main/java/Facilities/SeattleFacility.java
@@ -16,8 +16,8 @@ import java.util.Map;
 public final class SeattleFacility implements Facility, Comparable<Facility> {
 
     private String location = "Seattle, WA";
-    private long ratePerDay = 8;
-    private long costPerDay = 300;
+    private Integer ratePerDay = 8;
+    private Integer costPerDay = 300;
     private ArrayList<FacilityEdge> neighbors = new ArrayList<>();
 
     private HashMap<Item, Integer> activeInventory = new HashMap<>();
@@ -58,7 +58,7 @@ public final class SeattleFacility implements Facility, Comparable<Facility> {
     }
 
     @Override
-    public Long getRatePerDay() {
+    public Integer getRatePerDay() {
         return ratePerDay;
     }
 
@@ -68,7 +68,7 @@ public final class SeattleFacility implements Facility, Comparable<Facility> {
     }
 
     @Override
-    public Long getCostPerDay() {
+    public Integer getCostPerDay() {
         return costPerDay;
     }
